requirements.txt
streamlit==1.29.0
langchain==0.0.348
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pypdf==3.17.0
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
tiktoken==0.7.0
openai==0.28.2